Support newer versions of TextExpander. (#12)
* Support TextExpander 3.

* Fix typo.

* Convert TE cursor token to Alfred cursor token.

This converts the `%|` token in TextExpander snippets—which means "put the cursor here after expansion"—to `{cursor}`, the Alfred equivalent.

Note that this will not work on newer versions of TextExpander (v6.5+) that use a visual editor for tokens and do not include these in the `plainText` key but only in the `json` key, which is serialized and not easily parsed.

*  Rescue plain text from non-plain-text snippets.

Newer versions of TextExpander store snippets that have rich text and/or tokens in a serialized JSON object. Rather than rejecting non-plain-text snippets outright, attempt to rescue the plain text from that object.

* Capture more from JSON-defined snippets.

* Never mind, the `plainText` key always gets created.

So we don't need to rescue text from the JSON after all. 🙈

* Simplify.

* Restore check for existence of plainText.